refactor registries location
