+++ packages/builder/package.json
@@ -27,6 +27,7 @@
     "@types/lodash": "^4.14.183",
     "@types/pako": "^2.0.0",
     "jest": "^29.5.0",
+    "promise-events": "^0.2.4",
     "rollup": "^3.19.1",
     "ts-jest": "^29.0.5"
   },
