+++ packages/builder/src/registry.ts
@@ -1,5 +1,6 @@
 import { ethers, Overrides } from 'ethers';
 import Debug from 'debug';
+import EventEmitter from 'promise-events';
 
 import CannonRegistryAbi from './abis/CannonRegistry';
 
@@ -20,6 +21,71 @@ export abstract class CannonRegistry {
     }
 
     return null;
+  }
+}
+
+/**
+ * keeps track of packages in a simple JS object
+ * useful for testing and deployment dry-runs
+ */
+export class InMemoryRegistry extends CannonRegistry {
+  readonly pkgs: { [name: string]: { [variant: string]: string } } = {};
+
+  count = 0;
+
+  async publish(packagesNames: string[], variant: string, url: string): Promise<string[]> {
+    const receipts: string[] = [];
+    for (const name of packagesNames) {
+      if (!this.pkgs[name]) {
+        this.pkgs[name] = {};
+      }
+
+      this.pkgs[name][variant] = url;
+      receipts.push((++this.count).toString());
+    }
+
+    return receipts;
+  }
+
+  async getUrl(packageRef: string, variant: string): Promise<string | null> {
+    const baseResolved = await super.getUrl(packageRef, variant);
+    if (baseResolved) {
+      return baseResolved;
+    }
+
+    return this.pkgs[packageRef][variant];
+  }
+}
+
+export class FallbackRegistry extends EventEmitter implements CannonRegistry {
+  readonly registries: any[];
+
+  constructor(registries: any[]) {
+    super();
+    this.registries = registries;
+  }
+
+  async getUrl(packageRef: string, variant: string): Promise<string | null> {
+    for (const registry of this.registries) {
+      try {
+        const result = await registry.getUrl(packageRef, variant);
+
+        if (result) {
+          await this.emit('getUrl', { packageRef, variant, result, registry });
+          return result;
+        }
+      } catch (err) {
+        debug('WARNING: error caught in registry:', err);
+      }
+    }
+
+    return null;
+  }
+
+  async publish(packagesNames: string[], variant: string, url: string): Promise<string[]> {
+    debug('publish to fallback database: ', packagesNames);
+    // the fallback registry is usually something easy to write to or get to later
+    return _.first(this.registries).publish(packagesNames, variant, url);
   }
 }
