+++ packages/cli/package.json
@@ -42,7 +42,6 @@
     "ethers": "^5.7.1",
     "fs-extra": "^10.1.0",
     "lodash": "^4.17.21",
-    "promise-events": "^0.2.4",
     "prompts": "^2.4.2",
     "semver": "^7.3.7",
     "table": "^6.8.0",
